feat: Add KlkComponent to app module and update home page component
The code changes include adding the `KlkComponent` to the `declarations` array in the `AppModule` and updating the `HomePageComponent` template to use the `pokemons$` observable instead of the `pokemons` array.

Based on the recent user commits, it seems that the developer has been working on adding the home page component and updating the routing module.

Following the established commit message conventions in the repository, the suggested commit message is "feat: Add KlkComponent to app module and update home page component".